feat: move transpiler to separate package and use as web worker
