+++ packages/transpiler/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "@bos-web-engine/transpiler",
+  "version": "1.0.0",
+  "description": "",
+  "main": "lib/index.js",
+  "scripts": {
+    "build": "tsc -p tsconfig.json",
+    "dev": "tsc -w -p tsconfig.json",
+    "lint": "eslint src/**/*.ts*",
+    "lint:fix": "eslint container --fix"
+  },
+  "dependencies": {
+    "@babel/standalone": "^7.22.14",
+    "@bos-web-engine/container": "workspace:*",
+    "@near-js/providers": "^0.0.7"
+  },
+  "devDependencies": {
+    "@near-js/types": "^0.0.4",
+    "@types/node": "^17.0.12",
+    "eslint": "^7.32.0",
+    "eslint-config-custom": "workspace:*",
+    "tsconfig": "workspace:*",
+    "typescript": "^4.5.2"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
+}
