+++ packages/transpiler/tsconfig.json
@@ -1,0 +1,26 @@
+{
+  "extends": "tsconfig/react-library.json",
+  "include": [
+    "src"
+  ],
+  "exclude": ["lib", "dist", "build", "node_modules"],
+  "compilerOptions": {
+    "outDir": "./lib",
+    "lib": [
+      "dom"
+    ]
+  },
+  "watchOptions": {
+    // Use native file system events for files and directories
+    "watchFile": "useFsEvents",
+    "watchDirectory": "useFsEvents",
+    // Poll files for updates more frequently
+    // when they're updated a lot.
+    "fallbackPolling": "dynamicPriority",
+    // Don't coalesce watch notification
+    "synchronousWatchDirectory": true,
+    // Finally, two additional settings for reducing the amount of possible
+    // files to track  work from these directories
+    "excludeDirectories": ["**/node_modules", "_build"]
+  }
+}
